+++ api/app/com/foreignlanguagereader/api/controller/v1/language/DefinitionController.scala
@@ -26,17 +26,6 @@ class DefinitionController @Inject() (
 ) extends BaseController {
   val logger: Logger = Logger(this.getClass)
 
-  implicit def pathBinder: PathBindable[Language] =
-    new PathBindable[Language] {
-      override def bind(key: String, value: String): Either[String, Language] =
-        fromString(value) match {
-          case Some(language) =>
-            Right(language)
-          case _ => Left(value)
-        }
-      override def unbind(key: String, value: Language): String = value.toString
-    }
-
   def definition(wordLanguage: Language, word: String): Action[AnyContent] =
     Action.async {
       getDefinition(wordLanguage, Language.ENGLISH, word)
