+++ api/app/com/foreignlanguagereader/api/controller/v1/language/DocumentController.scala
@@ -24,17 +24,6 @@ class DocumentController @Inject() (
 ) extends BaseController {
   val logger: Logger = Logger(this.getClass)
 
-  implicit def pathBinder: PathBindable[Language] =
-    new PathBindable[Language] {
-      override def bind(key: String, value: String): Either[String, Language] =
-        fromString(value) match {
-          case Some(language) =>
-            Right(language)
-          case _ => Left(value)
-        }
-      override def unbind(key: String, value: Language): String = value.toString
-    }
-
   implicit val documentRequestReader: Reads[DocumentRequest] =
     (JsPath \ "text").read[String].map(text => new DocumentRequest(text))
